chore(tests): setup testing application in maximized mode with dettached devtools
